libs/bpm-ai-core/pyproject.toml
[tool.poetry]
name = "bpm-ai-core"
version = "0.1.1"
description = ""
authors = ["Benjoyo <bennet.krause@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = {version = "^1.0.0b3", allow-prereleases = true}
pytest = "^7.4.2"
pytest-dotenv = "^0.5.2"
jinja2 = "^3.1.2"
tenacity = "^8.2.3"
langsmith = "^0.0.54"
pydantic = "^2.4.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
